pyproject.toml
[project]
name = "problem-set-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pyyaml>=6.0.2",
]
